% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_enriched.R
\name{bind_enriched}
\alias{bind_enriched}
\title{Bind multiple enriched_refs objects}
\usage{
bind_enriched(...)
}
\arguments{
\item{...}{Objects of class \code{enriched_refs}.}
}
\value{
A combined \code{enriched_refs} object.
}
\description{
Bind multiple enriched_refs objects
}
