% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_enriched_bib.R
\name{build_bib_records}
\alias{build_bib_records}
\title{Build BibTeX-like records from an enriched_refs object}
\usage{
build_bib_records(x)
}
\arguments{
\item{x}{An \code{enriched_refs} object.}
}
\value{
A list of records ready for Python BibTeX writing.
}
\description{
Build BibTeX-like records from an enriched_refs object
}
\keyword{internal}
