% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_openalex.R
\name{build_openalex_abstract}
\alias{build_openalex_abstract}
\title{Rebuild abstract from OpenAlex inverted index}
\usage{
build_openalex_abstract(inverted_index)
}
\arguments{
\item{inverted_index}{OpenAlex abstract inverted index.}
}
\value{
Character scalar or NA.
}
\description{
Rebuild abstract from OpenAlex inverted index
}
\keyword{internal}
