% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidates.R
\name{candidates}
\alias{candidates}
\title{Extract candidate matches}
\usage{
candidates(x)
}
\arguments{
\item{x}{An \code{enriched_refs} object.}
}
\value{
A tibble of candidate matches.
}
\description{
Extract candidate matches
}
