% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_enriched_bib.R
\name{choose_abstract_value}
\alias{choose_abstract_value}
\title{Choose the preferred abstract value for export}
\usage{
choose_abstract_value(row)
}
\arguments{
\item{row}{One-row tibble.}
}
\value{
Character scalar or NA.
}
\description{
Choose the preferred abstract value for export
}
\keyword{internal}
