% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_crossref.R
\name{clean_html_tags}
\alias{clean_html_tags}
\title{Remove simple HTML/XML tags from text}
\usage{
clean_html_tags(text)
}
\arguments{
\item{text}{Character scalar.}
}
\value{
Cleaned character scalar.
}
\description{
Remove simple HTML/XML tags from text
}
\keyword{internal}
