% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_backend.R
\name{define_python_bib_helpers}
\alias{define_python_bib_helpers}
\title{Define Python helpers for BibTeX writing}
\usage{
define_python_bib_helpers()
}
\value{
Invisibly TRUE.
}
\description{
Define Python helpers for BibTeX writing
}
\keyword{internal}
