% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_backend.R
\name{define_python_ris_helpers}
\alias{define_python_ris_helpers}
\title{Define Python helpers for RIS parsing}
\usage{
define_python_ris_helpers()
}
\value{
Invisibly TRUE.
}
\description{
Define Python helpers for RIS parsing
}
\keyword{internal}
