% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_backend.R
\name{ensure_python_rispy}
\alias{ensure_python_rispy}
\title{Ensure rispy is available in the active Python environment}
\usage{
ensure_python_rispy()
}
\value{
Invisibly TRUE.
}
\description{
Ensure rispy is available in the active Python environment
}
\keyword{internal}
