% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_enriched_bib.R
\name{extract_note_from_raw}
\alias{extract_note_from_raw}
\title{Extract note from raw parsed RIS record}
\usage{
extract_note_from_raw(raw)
}
\arguments{
\item{raw}{One raw record.}
}
\value{
Character scalar or NA.
}
\description{
Extract note from raw parsed RIS record
}
\keyword{internal}
