% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_openaire.R
\name{extract_openaire_abstract}
\alias{extract_openaire_abstract}
\title{Extract abstract from an OpenAIRE item}
\usage{
extract_openaire_abstract(item)
}
\arguments{
\item{item}{A single OpenAIRE item.}
}
\value{
Character scalar or NA.
}
\description{
Extract abstract from an OpenAIRE item
}
\keyword{internal}
