% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_openaire.R
\name{extract_openaire_doi}
\alias{extract_openaire_doi}
\title{Extract DOI from an OpenAIRE item}
\usage{
extract_openaire_doi(item)
}
\arguments{
\item{item}{A single OpenAIRE item.}
}
\value{
Character scalar or NA.
}
\description{
Extract DOI from an OpenAIRE item
}
\keyword{internal}
