% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_enriched_bib.R
\name{first_non_missing}
\alias{first_non_missing}
\title{Return first non-missing value from arguments}
\usage{
first_non_missing(...)
}
\arguments{
\item{...}{Candidate values.}
}
\value{
First non-missing value, or NA.
}
\description{
Return first non-missing value from arguments
}
\keyword{internal}
