% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_abstract_qc.R
\name{flag_abstract_quality}
\alias{flag_abstract_quality}
\title{Flag abstract quality}
\usage{
flag_abstract_quality(x)
}
\arguments{
\item{x}{Character vector of abstracts.}
}
\value{
Character vector of abstract-quality labels.
}
\description{
Flag abstract quality
}
\keyword{internal}
