% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_qc.R
\name{flag_title_quality}
\alias{flag_title_quality}
\title{Flag suspicious title strings}
\usage{
flag_title_quality(x)
}
\arguments{
\item{x}{Character vector of titles.}
}
\value{
Character vector of quality labels.
}
\description{
Flag suspicious title strings
}
\keyword{internal}
