% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_enriched_bib.R
\name{is_missing_value}
\alias{is_missing_value}
\title{Check whether a value should be treated as missing}
\usage{
is_missing_value(x)
}
\arguments{
\item{x}{Any scalar.}
}
\value{
Logical scalar.
}
\description{
Check whether a value should be treated as missing
}
\keyword{internal}
