% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_enriched_bib.R
\name{make_bib_key}
\alias{make_bib_key}
\title{Make a BibTeX key}
\usage{
make_bib_key(author, year, title, record_id)
}
\arguments{
\item{author}{Author string.}

\item{year}{Year string/integer.}

\item{title}{Title string.}

\item{record_id}{Integer record id.}
}
\value{
Character scalar.
}
\description{
Make a BibTeX key
}
\keyword{internal}
