% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_enriched_bib.R
\name{map_ris_type_to_bib}
\alias{map_ris_type_to_bib}
\title{Map RIS type to a BibTeX entry type}
\usage{
map_ris_type_to_bib(x)
}
\arguments{
\item{x}{RIS type string.}
}
\value{
BibTeX entry type.
}
\description{
Map RIS type to a BibTeX entry type
}
\keyword{internal}
