% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{normalize_text}
\alias{normalize_text}
\title{Normalize text for matching}
\usage{
normalize_text(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Character vector.
}
\description{
Normalize text for matching
}
\keyword{internal}
