% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_enriched_bib.R
\name{null_if_missing}
\alias{null_if_missing}
\title{Convert missing values to NULL for Python export}
\usage{
null_if_missing(x)
}
\arguments{
\item{x}{Any scalar.}
}
\value{
NULL or scalar.
}
\description{
Convert missing values to NULL for Python export
}
\keyword{internal}
