% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{provenance}
\alias{provenance}
\title{Extract provenance information}
\usage{
provenance(x)
}
\arguments{
\item{x}{An \code{enriched_refs} object.}
}
\value{
A tibble of field-level provenance.
}
\description{
Extract provenance information
}
