% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_crossref.R
\name{query_crossref}
\alias{query_crossref}
\title{Query Crossref for candidate metadata}
\usage{
query_crossref(records)
}
\arguments{
\item{records}{A tibble with at least record_id and title.}
}
\value{
A tibble of candidate matches.
}
\description{
Query Crossref for candidate metadata
}
