% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input.R
\name{read_input_refs}
\alias{read_input_refs}
\title{Read and normalize input references}
\usage{
read_input_refs(
  input,
  input_type = c("auto", "ris", "bib", "csv", "data.frame", "titles")
)
}
\arguments{
\item{input}{File path, tibble, data.frame, or character vector.}

\item{input_type}{Input type.}
}
\value{
A tibble with standardized input columns.
}
\description{
Read and normalize input references
}
