% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_candidates.R
\name{score_candidates}
\alias{score_candidates}
\title{Score candidate matches}
\usage{
score_candidates(
  candidates,
  title_weight = 0.85,
  id_weight = 0.05,
  author_weight = 0.05,
  year_weight = 0.03,
  journal_weight = 0.02
)
}
\arguments{
\item{candidates}{A tibble of candidates.}

\item{title_weight}{Weight for title score.}

\item{id_weight}{Weight for identifier score.}

\item{author_weight}{Weight for author score.}

\item{year_weight}{Weight for year score.}

\item{journal_weight}{Weight for journal score.}
}
\value{
The input tibble with a confidence_score column.
}
\description{
Score candidate matches
}
