% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_summary.R
\name{summary.enriched_refs}
\alias{summary.enriched_refs}
\title{Summarize an enriched_refs object}
\usage{
\method{summary}{enriched_refs}(object, ...)
}
\arguments{
\item{object}{An \code{enriched_refs} object.}

\item{...}{Unused.}
}
\value{
A summary list, invisibly.
}
\description{
Summarize an enriched_refs object
}
