% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{title_similarity}
\alias{title_similarity}
\title{Compute title similarity}
\usage{
title_similarity(x, y)
}
\arguments{
\item{x}{First title.}

\item{y}{Second title.}
}
\value{
Numeric score between 0 and 1.
}
\description{
Compute title similarity
}
\keyword{internal}
