% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_enriched_bib.R
\name{write_enriched_bib}
\alias{write_enriched_bib}
\title{Write enriched references to a BibTeX file}
\usage{
write_enriched_bib(x, path)
}
\arguments{
\item{x}{An \code{enriched_refs} object.}

\item{path}{Output \code{.bib} file path.}
}
\value{
Invisibly returns \code{path}.
}
\description{
Write enriched references to a BibTeX file
}
